+++ src/screens/components/FrequencyInput.jsx
@@ -2,7 +2,16 @@ import React, { useCallback, useEffect, useState } from 'react'
 
 import ThemedTextInput from './ThemedTextInput'
 
+import debounce from 'debounce'
+
 const REMOVE_NON_DIGITS_REGEX = /[^0-9.]/g
+
+function reportChange ({ text, onChange, onChangeText, fieldId }) {
+  onChangeText && onChangeText(text)
+  onChange && onChange({ nativeEvent: { text }, fieldId })
+}
+
+const debouncedReportChange = debounce(reportChange, 1500)
 
 export default function FrequencyInput (props) {
   const { value, styles, textStyle, onChange, onChangeText, fieldId } = props
@@ -16,7 +25,8 @@ export default function FrequencyInput (props) {
     let { text } = event.nativeEvent
     text = text.replace(REMOVE_NON_DIGITS_REGEX, '')
     setInnerValue(text)
-  }, [])
+    debouncedReportChange({ text, onChange, onChangeText, fieldId })
+  }, [fieldId, onChange, onChangeText])
 
   const handleBlur = useCallback((event) => {
     const newEvent = { nativeEvent: { text: innerValue, target: event?.nativeEvent?.target } }